# Conflicts:
#	drivers/clock_control/clock_stm32_ll_common.c
#	drivers/gpio/gpio_renesas_ra_ioport.c
#	drivers/modem/modem_cellular.c
#	include/zephyr/rtio/rtio.h
#	subsys/net/ip/icmp.c
#	subsys/net/ip/net_if.c
#	subsys/net/ip/net_pkt.c
#	subsys/net/l2/ethernet/ethernet.c

+++ kernel/sched.c
@@ -1104,7 +1104,7 @@ int32_t z_impl_k_sleep(k_timeout_t timeout) {
     ticks = z_tick_sleep(ticks);
 
     int32_t ret = K_TIMEOUT_EQ(timeout, K_FOREVER) ? K_TICKS_FOREVER :
-                  k_ticks_to_ms_ceil64(ticks);
+                  (int32_t)k_ticks_to_ms_ceil64(ticks);
 
     SYS_PORT_TRACING_FUNC_EXIT(k_thread, sleep, timeout, ret);
